Create Week 1- Q3.cpp
Given an already sorted array of positive integers, design an algorithm and implement it using a program to find whether a given key element is present in the sorted array or not. For an array 
arr[n], search at the indexes arr[0], arr[2], arr[4],.....,arr[2k] and so on. Once the interval (arr[2k] < key < arr[ 2k+1] ) is found, perform a linear search operation from the index 2k
to find the element key. (Complexity < O(n), where n is the number of elements need to be scanned for searching):
#Jumpsearch